glass: move the IP/FQDN validator
Move the IP/FQDN validator to a more centralized
place so that it can be used by other components
as well.
