+++ src/glass/src/app/pages/register-page/register-page.component.html
@@ -27,7 +27,7 @@
                     translate>
                 This field is required.
               </span>
-              <span *ngIf="formGroup.hasError('address', 'address')"
+              <span *ngIf="formGroup.hasError('hostAddress', 'address')"
                     translate>
                 This field must be an IP address or FQDN (&lt;address&gt;:&lt;port&gt;).
               </span>
